requirements.txt
python-telegram-bot[job-queue]==21.6
yookassa==3.4.0
git+https://github.com/Neveix/tg_bot_base.git@v1.3.17